daemon/pyproject.toml
[tool.poetry]
name = "core"
version = "9.0.3"
description = "CORE Common Open Research Emulator"
authors = ["Boeing Research and Technology"]
license = "BSD-2-Clause"
repository = "https://github.com/coreemu/core"
documentation = "https://coreemu.github.io/core/"
include = [
    "core/api/grpc/*",
    "core/configservices/*/templates",
    "core/constants.py",
    "core/gui/data/**/*",
]
exclude = ["core/constants.py.in"]

[tool.poetry.scripts]
core-daemon = "core.scripts.daemon:main"
core-cli = "core.scripts.cli:main"
core-gui = "core.scripts.gui:main"
core-player = "core.scripts.player:main"
core-route-monitor = "core.scripts.routemonitor:main"
core-service-update = "core.scripts.serviceupdate:main"
core-cleanup = "core.scripts.cleanup:main"

[tool.poetry.dependencies]
python = "^3.9"
fabric = "2.7.1"
grpcio = "1.54.2"
invoke = "1.7.3"
lxml = "4.9.1"
netaddr = "0.7.19"
protobuf = "4.21.9"
pyproj = "3.3.1"
Pillow = "9.4.0"
Mako = "1.2.3"
PyYAML = "6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "6.2.5"
grpcio-tools = "1.54.2"
black = "22.12.0"
flake8 = "3.8.2"
isort = "4.3.21"
mock = "4.0.2"
pre-commit = "2.1.1"

[tool.isort]
skip_glob = "*_pb2*.py,doc,build"
multi_line_output = 3
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
line_length = 88

[tool.black]
line_length = 88
exclude = ".+_pb2.*.py|doc/|build/|__pycache__/"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

